adjust LocationSchema and reinstall realm packages
