+++ package.json
@@ -28,9 +28,9 @@
   "dependencies": {
     "@exponent/ex-navigation": "latest",
     "octicons": "^5.0.1",
-    "react": "~15.4.0-rc.4",
-    "react-native": "0.38.0",
-    "react-native-maps": "^0.11.0",
+    "react": "15.4.1",
+    "react-native": "0.39.2",
+    "react-native-maps": "^0.12.2",
     "react-native-svg": "^4.4.0",
     "react-native-vector-icons": "^3.0.0",
     "react-redux": "^3.1.0",
