+++ src/realm/location/methods.js
@@ -1,0 +1,9 @@
+import realm from '../index'
+
+export const saveLocations = (locations) => {
+    realm.write(() => {
+        locations.forEach((location) => {
+            realm.create('Location', location, true)
+        })  
+    })
+}
