+++ src/realm/location/schema.js
@@ -1,0 +1,11 @@
+const LocationSchema = {
+    name: 'Location',
+    primaryKey: 'locationId',
+    properties: {
+        locationId: 'string',
+        lat: 'string',
+        lng: 'string',
+    },
+}
+
+export default LocationSchema
